refactor: organize gesture modules
