+++ packages/react-native-screen-transitions/src/shared/providers/screen/gestures/pan/behavior/pan-reset.ts
@@ -1,11 +1,11 @@
-import { EPSILON, FALSE, TRUE } from "../../../../constants";
-import type { GestureStoreMap } from "../../../../stores/gesture.store";
-import type { AnimationConfig } from "../../../../types/animation.types";
+import { EPSILON, FALSE, TRUE } from "../../../../../constants";
+import type { GestureStoreMap } from "../../../../../stores/gesture.store";
+import type { AnimationConfig } from "../../../../../types/animation.types";
 import {
 	animateResetValue,
 	clearGestureSettlingIfResting,
 	getGestureResetSpec,
-} from "../shared/reset";
+} from "../../shared/reset";
 
 interface ResetPanGestureValuesProps {
 	spec?: AnimationConfig;
