+++ packages/react-native-screen-transitions/src/shared/providers/screen/gestures/pinch/activation/use-pinch-activation.ts
@@ -4,12 +4,12 @@ import type {
 	GestureTouchEvent,
 } from "react-native-gesture-handler";
 import type { SharedValue } from "react-native-reanimated";
-import type { ScreenOptionsContextValue } from "../../options";
-import { resolvePinchRuntime } from "../shared/runtime";
+import type { ScreenOptionsContextValue } from "../../../options";
+import { resolvePinchRuntime } from "../../shared/runtime";
 import type {
 	GestureCompositionActivation,
 	PinchGestureRuntime,
-} from "../types";
+} from "../../types";
 
 interface UsePinchActivationProps {
 	runtime: SharedValue<PinchGestureRuntime>;
