+++ packages/react-native-screen-transitions/src/shared/providers/screen/gestures/rotation/behavior/rotation-lifecycle.ts
@@ -1,4 +1,4 @@
-import type { RotationGestureEvent, RotationGestureRuntime } from "../types";
+import type { RotationGestureEvent, RotationGestureRuntime } from "../../types";
 
 export const startRotationGesture = (
 	gestures: RotationGestureRuntime["stores"]["gestures"],
